+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 - Set icon for menuitem by status of the pod
-- Add key namespace to settings.json
+- Add namespace to settings.json
 - Bugfix: open from the system tray sometimes did not work
+- Add code_command to settings.json
 
 ## v0.2.0 (2024-12-16)